Allow the option for "real" IBD
If IBD_PEER_ADDRESS is left empty, interpret that to mean that we want to
do an actual IBD from the P2P network.

Though an imprecise measure, I think this is an interesting thing to keep tabs
on. In any case, it's practical: we currently don't have enough machines to
provide dedicated fully synced, local peers for each benchmark runner we have.