Update default pr-bench chain region to post-segwit
Adds multiple pre-built pruned datadirs and defaults to one starting
roughly a year ago (height 667_200), which is more accurate of
contemporary validation since it includes a higher proportion of segwit
spends.

This also introduces a framework for more easily adding and articulating
the use of chain regions to benchmark against.